Move Playwright tests to separate class
